🐛 Fix: 버그 수정

+++ app/src/main/java/com/example/foodfood/api/userInfo/Data.kt
@@ -2,7 +2,7 @@ package com.example.foodfood.api.userInfo
 
 data class Data(
     val age: Int,
-    val height: Double,
+    val height: Int,
     val name: String,
-    val weight: Double
+    val weight: Int
 )
